Fixed Tables & Test Scripts
Fixed the tables not being added into database and causing an error.
The solution was to add each table manually with separate execute functions.

Created several test scripts to test the frame functionality as well as a dropdown
menu displaying different frames via forgetting the grid of the current row.
Whether or not we will use this feature is up to the group or when the
remaining time is too short to implement this feature.

Figured out how grid and pack works and why the TclError: "cannot use
geometry manager grid inside ..." happens, and fixed the issue in the test script.

I will need to figure out how to populate the tables with data. Fortunately, we
have the lecture stuff to help up insert the data into the tables, but not import
the data from the .csv files.