Functionality For Query Button
Created functionality for each applicable query which changes the results of each query.
Currently, there is no actual computation with each query, but it is a small commit that is worth committing.

Additionally, I made the query results be changed by configuring the commit and moved it outside of the function
so that it can be accessed statically.