Completed Part 2 - Query 1 Insert In Script
I created a function to do the insertion for Part 2 - Query 1 (see *Query Functions* section)
I included a verification for the case that the user does not include all of the required information to be inserted into the database.
I also included the command to clear each entry.

I readjusted the names of some labels.
I added the appropriate text to each corresponding label.