Query Functions Section Update
Updated the Query Function section to help us start working on the separate queries.
We still need to work on how to use the entries and include them in the Dropdown Command section.
We should not need to touch the Complete Queries Command section as everything is already included in the stuff above.